Refactored patch for CVE-2007-4559
