+++ seisbench/data/neic.py
@@ -109,29 +109,7 @@ class NEIC(BenchmarkDataset):
         tar_files = [x for x in path_original.iterdir() if x.name.endswith(".tar.gz")]
         for tar_path in tar_files:
             with tarfile.open(tar_path, "r:gz") as file:
-                
-                import os
-                
-                def is_within_directory(directory, target):
-                    
-                    abs_directory = os.path.abspath(directory)
-                    abs_target = os.path.abspath(target)
-                
-                    prefix = os.path.commonprefix([abs_directory, abs_target])
-                    
-                    return prefix == abs_directory
-                
-                def safe_extract(tar, path=".", members=None, *, numeric_owner=False):
-                
-                    for member in tar.getmembers():
-                        member_path = os.path.join(path, member.name)
-                        if not is_within_directory(path, member_path):
-                            raise Exception("Attempted Path Traversal in Tar File")
-                
-                    tar.extractall(path, members, numeric_owner=numeric_owner) 
-                    
-                
-                safe_extract(file, path_unpacked)
+                seisbench.util.safe_extract_tar(file, path_unpacked)
 
         groups = [("P", "Train"), ("S", "Train"), ("P", "Test"), ("S", "Test")]
